django-ocr-api/pyproject.toml
[project]
name = "django-rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "dvc>=3.59.0",
    "dvc-s3>=3.2.0",
    "onnxruntime>=1.20.1",
    "paddleocr>=2.9.1",
    "paddlepaddle>=2.6.2",
    "pillow>=11.1.0",
]
